Refactor api error handler
